requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.7.14
cffi==1.17.1
chainlit==2.6.2
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.115.14
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
gritql==0.2.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-cli==0.0.36
langchain-community==0.3.27
langchain-core==0.3.69
langchain-huggingface==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langserve==0.3.1
langsmith==0.4.8
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
openai==1.97.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.42.0
opentelemetry-instrumentation-anthropic==0.42.0
opentelemetry-instrumentation-bedrock==0.42.0
opentelemetry-instrumentation-chromadb==0.42.0
opentelemetry-instrumentation-cohere==0.42.0
opentelemetry-instrumentation-crewai==0.42.0
opentelemetry-instrumentation-google-generativeai==0.42.0
opentelemetry-instrumentation-groq==0.42.0
opentelemetry-instrumentation-haystack==0.42.0
opentelemetry-instrumentation-lancedb==0.42.0
opentelemetry-instrumentation-langchain==0.42.0
opentelemetry-instrumentation-llamaindex==0.42.0
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.42.0
opentelemetry-instrumentation-mcp==0.42.0
opentelemetry-instrumentation-milvus==0.42.0
opentelemetry-instrumentation-mistralai==0.42.0
opentelemetry-instrumentation-ollama==0.42.0
opentelemetry-instrumentation-openai==0.42.0
opentelemetry-instrumentation-openai-agents==0.42.0
opentelemetry-instrumentation-pinecone==0.42.0
opentelemetry-instrumentation-qdrant==0.42.0
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.42.0
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.42.0
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.42.0
opentelemetry-instrumentation-transformers==0.42.0
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.42.0
opentelemetry-instrumentation-watsonx==0.42.0
opentelemetry-instrumentation-weaviate==0.42.0
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.11
opentelemetry-util-http==0.55b1
orjson==3.11.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
sentry-sdk==2.33.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==1.8.2
starlette==0.46.2
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
traceloop-sdk==0.42.0
transformers==4.53.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==0.20.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
